% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors}
\alias{colors}
\title{CDS Colo(u)r Palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
colors
}
\description{
A list of color vectors that makes using CDS color palettes easier.
}
\examples{
\dontrun{
  scale_color_manual(values = cds::colors$missions$dark)
  scale_fill_manual(values = cds::colors$primary[c(1,2,7)])
}
}
\keyword{datasets}
