% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colours}
\alias{colours}
\title{CDS Colo(u)r Palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
colours
}
\description{
A list of colour vectors that makes using CDS colour palettes easier.
This is exactly the same as cds::colors
}
\examples{
\dontrun{
  scale_colour_manual(values = cds::colors$missions$dark)
  scale_fill_manual(values = cds::colors$primary[c(1,2,7)])
}
}
\keyword{datasets}
