% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{.onLoad}
\alias{.onLoad}
\title{On Load}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{character string giving the library directory where the package defining the namespace was found.}

\item{pkgname}{character string giving the name of the package.}
}
\value{
TRUE if successful
}
\description{
On Load
}
