% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-name.R
\name{resolve_name}
\alias{resolve_name}
\title{Resolve a name to an official organisation name}
\usage{
resolve_name(x, lang = "en", warn = TRUE)
}
\arguments{
\item{x}{a character vector of identifiers}

\item{lang}{language to return organisations in, one of 'en', 'fr', or 'id' (for CDS-specific unique department IDs)}

\item{warn}{warn when an identifier can't be resolved}
}
\value{
a character vector of official organisation names
}
\description{
Resolve a name to an official organisation name
}
\examples{

resolve_name("IRCC")
# "Immigration, Refugees and Citizenship Canada"

resolve_name(c("NRCan", "ESDC"), lang = "fr")
# "Ressources naturelles Canada", "Emploi et Développement social Canada"

resolve_name("Transport Canada / Transports Canada", lang = "id")
# "001I9000004Vt6JIAS"

resolve_name("Department of Unicorns", lang = "en")
# NA with warning

}
