% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_cds}
\alias{theme_cds}
\title{CDS ggplot2 Theme}
\usage{
theme_cds()
}
\value{
theme() object that can be applied to ggplot objects
}
\description{
CDS ggplot2 Theme
}
\examples{
library(ggplot2)
my_plot <- ggplot(mpg, aes(x = hwy, y = cty))
my_plot <- my_plot + theme_bw()  # Can layer multiple themes
my_plot <- my_plot + theme_cds()

}
